[Refactor] 포인트 내역 생성 로직 분리
