+++ src/main/java/com/nhnacademy/heukbaekbookshop/point/history/controller/PointHistoryController.java
@@ -1,13 +1,10 @@
 package com.nhnacademy.heukbaekbookshop.point.history.controller;
 
-import com.nhnacademy.heukbaekbookshop.point.history.dto.request.PointHistoryRequest;
 import com.nhnacademy.heukbaekbookshop.point.history.dto.response.PointHistoryResponse;
 import com.nhnacademy.heukbaekbookshop.point.history.service.PointHistoryService;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -33,15 +30,5 @@ public class PointHistoryController {
     public ResponseEntity<BigDecimal> getCurrentBalanceByMemberId(@RequestHeader("X-USER-ID") Long memberId) {
         BigDecimal currentBalanceByCustomerId = pointHistoryService.getCurrentBalanceByCustomerId(memberId);
         return ResponseEntity.ok(currentBalanceByCustomerId);
-    }
-
-    @PostMapping("/histories")
-    public ResponseEntity<PointHistoryResponse> createPointHistory(
-            @RequestHeader("X-USER-ID") Long memberId,
-            @Valid @RequestBody PointHistoryRequest request
-    ) {
-        PointHistoryResponse pointHistory = pointHistoryService.createPointHistory(memberId, request);
-
-        return ResponseEntity.status(HttpStatus.CREATED).body(pointHistory);
     }
 }
