+++ src/main/java/com/nhnacademy/heukbaekbookshop/point/history/service/PointHistoryService.java
@@ -1,6 +1,5 @@
 package com.nhnacademy.heukbaekbookshop.point.history.service;
 
-import com.nhnacademy.heukbaekbookshop.point.history.dto.request.PointHistoryRequest;
 import com.nhnacademy.heukbaekbookshop.point.history.dto.response.PointHistoryResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -11,6 +10,4 @@ public interface PointHistoryService {
     Page<PointHistoryResponse> getPointHistoriesByCustomerId(Long customerId, Pageable pageable);
 
     BigDecimal getCurrentBalanceByCustomerId(Long customerId);
-
-    PointHistoryResponse createPointHistory(Long customerId, PointHistoryRequest pointHistoryRequest);
 }
